Move AnyIPEndpoint to a separate file & implement CustomStringConvertible on IPEndpoint<T>
